% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music2Param-class.R
\name{music2Param}
\alias{music2Param}
\title{Make an object of class music2Param}
\usage{
music2Param(
  bulkExpression = NULL,
  bulkExpressionIndependent = NULL,
  referenceExpression = NULL,
  cellScaleFactors = NULL,
  singleCellExpressionSet = NULL,
  singleCellExperiment = NULL,
  assayName = "counts",
  batchVariable = "SubjectName",
  cellTypeVariable = "cellType",
  conditionVariable = "experimentCondition",
  controlLabel = "control",
  caseLabel = "case",
  methodType = "MuSiC2",
  returnInfo = FALSE
)
}
\arguments{
\item{bulkExpression}{Bulk mixed signals matrix of samples, which can be 
matched to single-cell samples.}

\item{bulkExpressionIndependent}{Bulk mixed signals matrix of independent 
samples, which should not overlap samples in y.}

\item{referenceExpression}{Signature matrix of cell type-specific signals. 
If not provided, can be computed from a provided ExpressionSet containing 
single-cell data.}

\item{cellScaleFactors}{Cell size factor transformations of length equal to 
the K cell types to deconvolve.}

\item{assayName}{Expression data type (e.g. counts, logcounts, tpm, etc.).}

\item{batchVariable}{Name of variable identifying the batches in 
SingleCellExpressionSet pData/coldata.}

\item{cellTypeVariable}{Name of cell type labels variable in 
SingleCellExpressionSet pData/coldata.}

\item{conditionVariable}{Name of variable in y.eset and 
SingleCellExpressionSet containing condition labels.}

\item{controlLabel}{Label of control condition samples in condition variable.}

\item{caseLabel}{Label of case condition samples in condition variable.}

\item{methodType}{Name of method source library to call for music2_prop; 
either "MuSiC" or "MuSiC2".}

\item{returnInfo}{Whether to return metadata and original method outputs 
with predicted proportions.}

\item{SingleCellExpressionSet}{ExpressionSet of single-cell expression.}

\item{SingleCellExperiment}{SingleCellExperiment of single-cell expression.}
}
\value{
Object of class \linkS4class{music2Param}.
}
\description{
Main function to make a new object of class \linkS4class{music2Param}, with
defaults for required arguments.
}
\seealso{
\linkS4class{musicParam}
}
